requirements.txt
nonebot2
nonebot-adapter-onebot
httpx~=0.23.0
requests~=2.28.1
loguru~=0.6.0
nonebot-plugin-addfriend
openai